Rename project bundle name
