+++ android/app/build.gradle
@@ -129,7 +129,7 @@ android {
     }
 
     defaultConfig {
-        applicationId "com.gobarberapp"
+        applicationId "com.gobarberapp.roni"
         minSdkVersion rootProject.ext.minSdkVersion
         targetSdkVersion rootProject.ext.targetSdkVersion
         versionCode 1
