+++ android/app/src/main/java/com/gobarberapp/roni/MainActivity.java
@@ -1,4 +1,4 @@
-package com.gobarberapp;
+package com.gobarberapp.roni;
 
 import com.facebook.react.ReactActivity;
 import android.os.Bundle;
