+++ android/app/src/main/java/com/gobarberapp/roni/SplashActivity.java
@@ -1,4 +1,4 @@
-package com.gobarberapp;
+package com.gobarberapp.roni;
 
 import android.content.Intent;
 import android.os.Bundle;
